+++ src/execution/plugins/script-execution-tools/custom-function-missing.test.ts
@@ -3,32 +3,35 @@ import spaceTrim from 'spacetrim';
 import { PromptTemplatePipeline } from '../../../classes/PromptTemplatePipeline';
 import { promptTemplatePipelineStringToJson } from '../../../conversion/promptTemplatePipelineStringToJson';
 import { PromptTemplatePipelineString } from '../../../types/PromptTemplatePipelineString';
+import { assertsExecutionSuccessful } from '../../assertsExecutionSuccessful';
 import { createPtpExecutor } from '../../createPtpExecutor';
 import { MockedEchoNaturalExecutionTools } from '../natural-execution-tools/mocked/MockedEchoNaturalExecutionTools';
 import { CallbackInterfaceTools } from '../user-interface-execution-tools/callback/CallbackInterfaceTools';
 import { JavascriptEvalExecutionTools } from './javascript/JavascriptEvalExecutionTools';
 
-describe('createPtpExecutor + executing scripts in ptp', () => {
+describe('createPtpExecutor + missing custom function', () => {
     const ptbJson = promptTemplatePipelineStringToJson(
         spaceTrim(`
-            # Sample prompt
+            # Custom functions
 
-            Show how to execute a script
+            Show how to use custom postprocessing functions
 
             -   PTBK VERSION 1.0.0
-            -   INPUT  PARAMETER {thing} Any thing to buy
+            -   INPUT  PARAMETER {yourName} Name of the hero
 
-            ## Execution
+            ## Question
 
-            -   EXECUTE SCRIPT
+            -   SIMPLE TEMPLATE
+            -   POSTPROCESSING addHello
 
-            \`\`\`javascript
-            thing.split('a').join('b')
+            \`\`\`markdown
+            {yourName} the Evangelist
             \`\`\`
 
-            -> {bhing}
+            -> {greeting}
          `) as PromptTemplatePipelineString,
     );
+
     const ptp = PromptTemplatePipeline.fromJson(ptbJson);
     const ptpExecutor = createPtpExecutor({
         ptp,
@@ -37,7 +40,13 @@ describe('createPtpExecutor + executing scripts in ptp', () => {
             script: [
                 new JavascriptEvalExecutionTools({
                     isVerbose: true,
-                    // Note: [🕎] Custom functions are tested elsewhere
+
+                    // Note: [🕎]
+                    functions: {
+                        addHelloWithTypo(value) {
+                            return `Hello ${value}`;
+                        },
+                    },
                 }),
             ],
             userInterface: new CallbackInterfaceTools({
@@ -52,26 +61,9 @@ describe('createPtpExecutor + executing scripts in ptp', () => {
         },
     });
 
-    it('should work when every INPUT  PARAMETER defined', () => {
-        expect(ptpExecutor({ thing: 'apple' }, () => {})).resolves.toMatchObject({
-            isSuccessful: true,
-            outputParameters: {
-                bhing: 'bpple',
-            },
-        });
-        expect(ptpExecutor({ thing: 'a cup of coffee' }, () => {})).resolves.toMatchObject({
-            isSuccessful: true,
-            outputParameters: {
-                bhing: 'b cup of coffee',
-            },
-        });
-    });
-
-    it('should fail when some INPUT  PARAMETER is missing', () => {
-        expect(ptpExecutor({}, () => {})).resolves.toMatchObject({
-            errors: [new Error(`Parameter {thing} is not defined`)],
-        });
-
-        expect(() => ptpExecutor({}, () => {})).rejects.toThrowError(/Parameter \{thing\} is not defined/);
+    it('should throw error when custom postprocessing function does not exist', () => {
+        expect(() =>
+            ptpExecutor({ yourName: 'Matthew' }, () => {}).then(assertsExecutionSuccessful),
+        ).rejects.toThrowError(/Function \{addHello\} is not defined/);
     });
 });